chore(Dashboard2): move bench only components to /bench directory
